Add sandbox runtime tests and helpers
