+++ Cargo.toml
@@ -2,7 +2,9 @@
 members = [
     "crates/bpf-api",
     "crates/bpf-core",
+    "crates/bpf-host",
     "crates/cli",
+    "crates/sandbox-runtime",
     "crates/agent-lite",
     "crates/policy-core",
     "crates/policy-compiler",
