+++ crates/cli/Cargo.toml
@@ -10,18 +10,13 @@ documentation = "https://docs.rs/qqrm-cargo-warden"
 readme = "../../README.md"
 
 [dependencies]
-anyhow = "1"
-aya = "0.13"
 clap = { version = "4", features = ["derive"] }
-libc = "0.2"
-libseccomp = "0.3"
-bpf-api = { package = "qqrm-bpf-api", version = "0.1.0", path = "../bpf-api" }
 policy-core = { package = "qqrm-policy-core", version = "0.1.0", path = "../policy-core" }
 qqrm-policy-compiler = { version = "0.1.0", path = "../policy-compiler" }
+sandbox-runtime = { package = "qqrm-sandbox-runtime", version = "0.1.0", path = "../sandbox-runtime" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 toml = "0.8"
-qqrm-agent-lite = { version = "0.1.0", path = "../agent-lite" }
 
 [dev-dependencies]
 serial_test = "3"
