+++ crates/cli/src/main.rs
@@ -8,9 +8,7 @@ use std::io::{self, BufRead, BufReader, Write};
 use std::path::Path;
 use std::process::{Command, ExitStatus, exit};
 
-mod sandbox;
-
-use sandbox::Sandbox;
+use sandbox_runtime::Sandbox;
 
 #[derive(Copy, Clone, Debug, ValueEnum)]
 enum CliMode {
@@ -379,25 +377,6 @@ fn run_in_sandbox(command: Command, isolation: &IsolationConfig) -> io::Result<E
     };
     shutdown_result?;
     Ok(status)
-}
-
-#[cfg(not(test))]
-pub(crate) fn apply_seccomp(deny: &[String]) -> io::Result<()> {
-    use libseccomp::{ScmpAction, ScmpFilterContext, ScmpSyscall};
-    let mut filter = ScmpFilterContext::new_filter(ScmpAction::Allow).map_err(io::Error::other)?;
-    for name in deny {
-        if let Ok(sys) = ScmpSyscall::from_name(name) {
-            filter
-                .add_rule(ScmpAction::Errno(libc::EPERM), sys)
-                .map_err(io::Error::other)?;
-        }
-    }
-    filter.load().map_err(io::Error::other)
-}
-
-#[cfg(test)]
-pub(crate) fn apply_seccomp(_deny: &[String]) -> io::Result<()> {
-    Ok(())
 }
 
 fn handle_status() -> io::Result<()> {
