+++ crates/sandbox-runtime/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "qqrm-sandbox-runtime"
+version = "0.1.0"
+edition = "2024"
+license = "MIT OR Apache-2.0"
+description = "Runtime for cargo-warden sandbox execution."
+repository = "https://github.com/qqrm/cargo-warden"
+homepage = "https://github.com/qqrm/cargo-warden"
+documentation = "https://docs.rs/qqrm-sandbox-runtime"
+
+[dependencies]
+anyhow = "1"
+aya = "0.13"
+bpf-api = { package = "qqrm-bpf-api", version = "0.1.0", path = "../bpf-api" }
+libc = "0.2"
+libseccomp = "0.3"
+qqrm-agent-lite = { version = "0.1.0", path = "../agent-lite" }
+qqrm-policy-compiler = { version = "0.1.0", path = "../policy-compiler" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+[dev-dependencies]
+serial_test = "3"
+tempfile = "3"
