+++ crates/sandbox-runtime/src/agent.rs
@@ -1,0 +1,48 @@
+use anyhow::Error as AnyhowError;
+use aya::maps::{MapData, ring_buf::RingBuf};
+use qqrm_agent_lite::{self, Config as AgentConfig, Shutdown, ShutdownHandle};
+use std::io;
+use std::path::PathBuf;
+use std::thread;
+use std::time::Duration;
+
+pub(crate) struct AgentHandle {
+    shutdown: ShutdownHandle,
+    thread: Option<thread::JoinHandle<Result<(), AnyhowError>>>,
+}
+
+impl AgentHandle {
+    pub(crate) fn stop(mut self) -> io::Result<()> {
+        self.shutdown.request();
+        if let Some(handle) = self.thread.take() {
+            match handle.join() {
+                Ok(result) => result.map_err(|err| io::Error::other(err.to_string())),
+                Err(err) => Err(io::Error::other(format!("agent thread panicked: {err:?}"))),
+            }
+        } else {
+            Ok(())
+        }
+    }
+}
+
+impl Drop for AgentHandle {
+    fn drop(&mut self) {
+        if let Some(handle) = self.thread.take() {
+            self.shutdown.request();
+            let _ = handle.join();
+        }
+    }
+}
+
+pub(crate) fn start_agent(ring: RingBuf<MapData>, events_path: PathBuf) -> io::Result<AgentHandle> {
+    let cfg = AgentConfig::default();
+    let (shutdown, signal) = Shutdown::new(Duration::from_millis(100));
+    let thread = thread::Builder::new()
+        .name("qqrm-agent-lite".into())
+        .spawn(move || qqrm_agent_lite::run_with_shutdown(ring, &events_path, cfg, signal))
+        .map_err(|err| io::Error::other(format!("failed to spawn agent thread: {err}")))?;
+    Ok(AgentHandle {
+        shutdown,
+        thread: Some(thread),
+    })
+}
